Chore/cleanup code (#135)

+++ src/main/java/se/isselab/HAnS/referencing/FeatureReferenceUtil.java
@@ -29,6 +29,7 @@ import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.psi.search.searches.ReferencesSearch;
 import com.intellij.util.Range;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 import se.isselab.HAnS.FeatureAnnotationToDelete;
 import se.isselab.HAnS.featureLocation.FeatureFileMapping;
 import se.isselab.HAnS.featureLocation.FeatureLocation;
@@ -41,7 +42,6 @@ import se.isselab.HAnS.fileAnnotation.psi.FileAnnotationFileReferences;
 import se.isselab.HAnS.metrics.calculators.FeatureTangling;
 
 import java.util.*;
-import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
@@ -57,11 +57,12 @@ public class FeatureReferenceUtil {
     private static Map<String, List<PsiReference>> mapToRenameWhenAdding = new HashMap<>();
     private static final Map<Document, Set<Integer>> mapToDeleteAfterDeletingWithCode = new HashMap<>(); // stores document->empty lines to remove after renaming, because otherwise it would corrupt the document
 
+    private FeatureReferenceUtil() {}
+
     public static String getLPQ(FeatureModelFeature feature, String newName) {
         if (lpq == null) {
             return setLPQ(feature, newName);
-        }
-        else {
+        } else {
             return lpq;
         }
     }
@@ -85,7 +86,9 @@ public class FeatureReferenceUtil {
         mapToDeleteAfterDeletingWithCode.clear();
     }
 
-    public static Map<Document, Set<Integer>> getMapToDeleteAfterDeletingWithCode() { return mapToDeleteAfterDeletingWithCode; }
+    public static Map<Document, Set<Integer>> getMapToDeleteAfterDeletingWithCode() {
+        return mapToDeleteAfterDeletingWithCode;
+    }
 
     public static Map<FeatureModelFeature, List<PsiReference>> getElementsToRename() {
         return mapToRename;
@@ -99,7 +102,9 @@ public class FeatureReferenceUtil {
         return mapToRenameWhenAdding;
     }
 
-    public static Map<FeatureModelFeature, ArrayList<Object>> getMapToDeleteWithCode() { return mapToDeleteWithCode; }
+    public static Map<FeatureModelFeature, ArrayList<Object>> getMapToDeleteWithCode() {
+        return mapToDeleteWithCode;
+    }
 
     private static String setLPQ(FeatureModelFeature feature, String newName) {
         PsiElement fileCopy = feature.getContainingFile().copy();
@@ -107,11 +112,11 @@ public class FeatureReferenceUtil {
         fileCopy.accept(new PsiRecursiveElementWalkingVisitor() {
             @Override
             public void visitElement(@NotNull PsiElement element) {
-                if (element instanceof FeatureModelFeature){
-                    if (((FeatureModelFeature) element).getLPQText().equals(feature.getLPQText())) {
-                        e[0] = (FeatureModelFeature) element;
-                    }
+                if (element instanceof FeatureModelFeature featureModelFeature &&
+                        featureModelFeature.getLPQText().equals(feature.getLPQText())) {
+                    e[0] = featureModelFeature;
                 }
+
                 super.visitElement(element);
             }
         });
@@ -122,7 +127,8 @@ public class FeatureReferenceUtil {
             e[0].getNode().replaceChild(featureNode, newKeyNode);
         }
         origin = feature;
-        return lpq = e[0].getLPQText();
+        lpq = e[0].getLPQText();
+        return lpq;
     }
 
     public static void rename() {
@@ -130,10 +136,7 @@ public class FeatureReferenceUtil {
 
         for (Map.Entry<FeatureModelFeature, List<PsiReference>> entry : toRename.entrySet()) {
             String newLPQ = entry.getKey().getLPQText();
-            System.out.println(newLPQ);
-            for (PsiReference reference:entry.getValue()) {
-                System.out.println(reference.getElement().getText());
-                System.out.println(reference.getElement().getContainingFile());
+            for (PsiReference reference : entry.getValue()) {
                 reference.handleElementRename(newLPQ);
             }
         }
@@ -141,36 +144,23 @@ public class FeatureReferenceUtil {
 
     public static void setElementsToRenameWhenRenaming(FeatureModelFeature element, String newName) {
         Map<FeatureModelFeature, List<PsiReference>> toRename = new HashMap<>();
-
-        System.out.println("Will rename");
-
         List<FeatureModelFeature> elementsToRename = getElementsToRenameWhenRenaming(element, newName);
 
-        for (FeatureModelFeature e : elementsToRename) {
-            System.out.println(e.getLPQText());
-            List<PsiReference> referencedElements = new ArrayList<>();
-            for (PsiReference reference : ReferencesSearch.search(e)) {
-                referencedElements.add(reference);
-            }
-            toRename.put(e, referencedElements);
-        }
-
-        mapToRename = toRename;
+        mapElementsToRename(toRename, elementsToRename);
     }
 
     private static List<FeatureModelFeature> getElementsToRenameWhenRenaming(FeatureModelFeature element, String newName) {
-        List<FeatureModelFeature> elementsToRename= new ArrayList<>();
+        List<FeatureModelFeature> elementsToRename = new ArrayList<>();
         element.getContainingFile().accept(new PsiRecursiveElementWalkingVisitor() {
             @Override
             public void visitElement(@NotNull PsiElement e) {
-                if (e instanceof FeatureModelFeature){
-                    if (((FeatureModelFeature) e).getLPQText().contains(newName)) {
-                        elementsToRename.add((FeatureModelFeature) e);
-                    }
-                    else if (((FeatureModelFeature) e).getLPQText().contains(Objects.requireNonNull(element.getNode().findChildByType(FeatureModelTypes.FEATURENAME)).getText())) {
-                        elementsToRename.add((FeatureModelFeature) e);
-                    }
+                if (e instanceof FeatureModelFeature featureModelFeature &&
+                        (featureModelFeature.getLPQText().contains(newName) ||
+                                featureModelFeature.getLPQText().contains(Objects.requireNonNull(element.getNode()
+                                        .findChildByType(FeatureModelTypes.FEATURENAME)).getText()))) {
+                    elementsToRename.add(featureModelFeature);
                 }
+
                 super.visitElement(e);
             }
         });
@@ -196,12 +186,12 @@ public class FeatureReferenceUtil {
     }
 
     private static List<FeatureModelFeature> getElementsToRenameWhenAdding(FeatureModelFeature element, String newName) {
-        List<FeatureModelFeature> elementsToRename= new ArrayList<>();
+        List<FeatureModelFeature> elementsToRename = new ArrayList<>();
         element.getContainingFile().accept(new PsiRecursiveElementWalkingVisitor() {
             @Override
             public void visitElement(@NotNull PsiElement e) {
-                if (e instanceof FeatureModelFeature && ((FeatureModelFeature) e).getLPQText().contains(newName)){
-                    elementsToRename.add((FeatureModelFeature) e);
+                if (e instanceof FeatureModelFeature featureModelFeature && (featureModelFeature).getLPQText().contains(newName)) {
+                    elementsToRename.add(featureModelFeature);
                 }
                 super.visitElement(e);
             }
@@ -214,7 +204,7 @@ public class FeatureReferenceUtil {
 
         for (Map.Entry<FeatureModelFeature, List<PsiReference>> entry : toDelete.entrySet()) {
 
-            for (PsiReference reference:entry.getValue()) {
+            for (PsiReference reference : entry.getValue()) {
                 InjectedLanguageManager injManager = InjectedLanguageManager.getInstance(entry.getKey().getProject());
 
                 PsiLanguageInjectionHost host = injManager.getInjectionHost(reference.getElement());
@@ -224,7 +214,7 @@ public class FeatureReferenceUtil {
                         Runnable r = () -> {
                             PsiElement parent = reference.getElement().getParent();
                             String parentPsi = parent.getText(); // e.g. [A, B ,  C]
-                            String[] array = parentPsi.replaceAll("\\[|]|\\s", "").split(","); // transform to list of features {A, B, C}
+                            String[] array = parentPsi.replaceAll("\\[|]|\\s", "").split(","); // transform to list of features A, B, C
 
                             ArrayList<String> featuresArr = new ArrayList<>(Arrays.asList(array)); // transform to arraylist
                             featuresArr.remove(reference.getElement().getText()); // remove our element from arraylist
@@ -280,21 +270,16 @@ public class FeatureReferenceUtil {
             WriteCommandAction.runWriteCommandAction(project, r);
         }
 
-        if (reference.getElement().getPrevSibling() == null && reference.getElement().getNextSibling() == null) { // if it's last feature mapped to file
-            if (reference.getElement().getParent().getPrevSibling() != null) { // if newline exists
-                if (reference.getElement().getParent().getPrevSibling().getPrevSibling() != null &&
-                        reference.getElement().getParent().getPrevSibling().getPrevSibling() instanceof FileAnnotationFileReferences) { // get file list
-                    Runnable r = () -> {
-                        reference.getElement().getParent().getPrevSibling().getPrevSibling().delete(); // delete all files
-                    };
-                    WriteCommandAction.runWriteCommandAction(project, r);
-                }
-            }
+        if (reference.getElement().getPrevSibling() == null && reference.getElement().getNextSibling() == null &&
+                reference.getElement().getParent().getPrevSibling() != null &&
+                reference.getElement().getParent().getPrevSibling().getPrevSibling() instanceof FileAnnotationFileReferences) { // get file list
+            Runnable r = () ->
+                    reference.getElement().getParent().getPrevSibling().getPrevSibling().delete(); // delete all files
+            WriteCommandAction.runWriteCommandAction(project, r);
         }
 
-        Runnable r = () -> {
-            reference.getElement().delete(); // remove element
-        };
+
+        Runnable r = () -> reference.getElement().delete(); // remove element
         WriteCommandAction.runWriteCommandAction(project, r);
     }
 
@@ -371,7 +356,7 @@ public class FeatureReferenceUtil {
     }
 
     public static void setMapToDeleteWithCode(FeatureModelFeature feature) {
-        List<PsiElement> elementsToDelete= new ArrayList<>();
+        List<PsiElement> elementsToDelete = new ArrayList<>();
         traverseFeatureWithChildren(feature, elementsToDelete, new ArrayList<>());
 
         Project projectInstance = feature.getProject();
@@ -408,12 +393,12 @@ public class FeatureReferenceUtil {
 
                     if (entry == null) {
                         entry = new HashSet<>();
-                        for (int i = block.getStartLine(); i < block.getEndLine()+1; i++) {
+                        for (int i = block.getStartLine(); i < block.getEndLine() + 1; i++) {
                             entry.add(i);
                         }
                         mapToDrop.put(fm.getMappedPathPairMappedBy(), entry);
                     } else {
-                        for (int i = block.getStartLine(); i < block.getEndLine()+1; i++) {
+                        for (int i = block.getStartLine(); i < block.getEndLine() + 1; i++) {
                             entry.add(i);
                         }
                     }
@@ -502,11 +487,11 @@ public class FeatureReferenceUtil {
 
     // returns set of all features tangled with FeatureModelFeature feature
     private static HashSet<FeatureModelFeature> getTangledFeatures(HashMap<FeatureModelFeature, HashSet<FeatureModelFeature>> tanglingMap, FeatureModelFeature feature) {
-        HashSet<FeatureModelFeature> tangledFeatures= new HashSet<>();
+        HashSet<FeatureModelFeature> tangledFeatures = new HashSet<>();
 
-        for (FeatureModelFeature key : tanglingMap.keySet()) {
-            if(key.getLPQText().equals(feature.getLPQText())) {
-                tangledFeatures = tanglingMap.get(key);
+        for (var entry : tanglingMap.entrySet()) {
+            if (entry.getKey().getLPQText().equals(feature.getLPQText())) {
+                tangledFeatures = entry.getValue();
             }
         }
 
@@ -554,7 +539,7 @@ public class FeatureReferenceUtil {
                 newElement.setMainAnnotationType(baseLocation.getMainAnnotationType()); // 6: main annotation type
 
                 if (tangledLocation.getTangledAnnotationType().equals(FeatureFileMapping.AnnotationType.CODE) &&
-                    baseLocation.getMainAnnotationType().equals(FeatureFileMapping.AnnotationType.CODE)) {
+                        baseLocation.getMainAnnotationType().equals(FeatureFileMapping.AnnotationType.CODE)) {
                     ArrayList<Integer> rangeResult = getRange(baseLocation.getStartLine(), baseLocation.getEndLine(), tangledLocation.getStartLine(), tangledLocation.getEndLine());
                     if (rangeResult != null) {
                         newElement.setStartLine(rangeResult.get(0)); // 2: start line
@@ -585,7 +570,7 @@ public class FeatureReferenceUtil {
 
     // if two features are tangled using code annotation,
     // finds line range that includes both FeatureA and FeatureB
-    private static ArrayList<Integer> getRange(int startA, int endA, int startB, int endB) {
+    private static @Nullable ArrayList<Integer> getRange(int startA, int endA, int startB, int endB) {
         ArrayList<Integer> result = new ArrayList<>();
         Range<Integer> rangeA = new Range<>(startA, endA);
         Range<Integer> rangeB = new Range<>(startB, endB);
@@ -605,7 +590,7 @@ public class FeatureReferenceUtil {
             result.add(endB);
             return result;
         }
-        if (rangeB.isWithin(startA, true)){
+        if (rangeB.isWithin(startA, true)) {
             result.add(startB);
             result.add(endA);
             return result;
@@ -616,7 +601,7 @@ public class FeatureReferenceUtil {
     public static void setElementsToDelete(FeatureModelFeature element) {
         Map<FeatureModelFeature, List<PsiReference>> toDelete = new HashMap<>();
 
-        List<PsiElement> elementsToDelete= new ArrayList<>();
+        List<PsiElement> elementsToDelete = new ArrayList<>();
         traverseFeatureWithChildren(element, elementsToDelete, new ArrayList<>());
         GlobalSearchScope scope = GlobalSearchScope.allScope(element.getProject());
 
@@ -637,8 +622,8 @@ public class FeatureReferenceUtil {
 
         Map<FeatureModelFeature, List<PsiReference>> toRename = new HashMap<>();
 
-        List<PsiElement> elementsToDelete= new ArrayList<>();
-        List<String> lpqToDelete= new ArrayList<>();
+        List<PsiElement> elementsToDelete = new ArrayList<>();
+        List<String> lpqToDelete = new ArrayList<>();
         traverseFeatureWithChildren(element, elementsToDelete, lpqToDelete);
 
         List<FeatureModelFeature> elementsToRename = new ArrayList<>();
@@ -646,6 +631,10 @@ public class FeatureReferenceUtil {
             elementsToRename.addAll(getElementsToRenameWhenDeleting(((FeatureModelFeature) feature), lpqToDelete));
         }
 
+        mapElementsToRename(toRename, elementsToRename);
+    }
+
+    private static void mapElementsToRename(Map<FeatureModelFeature, List<PsiReference>> toRename, List<FeatureModelFeature> elementsToRename) {
         for (FeatureModelFeature e : elementsToRename) {
             List<PsiReference> referencedElements = new ArrayList<>();
             for (PsiReference reference : ReferencesSearch.search(e)) {
@@ -657,16 +646,16 @@ public class FeatureReferenceUtil {
     }
 
     private static List<FeatureModelFeature> getElementsToRenameWhenDeleting(FeatureModelFeature element, List<String> lpqToRemove) {
-        List<FeatureModelFeature> elementsToRename= new ArrayList<>();
+        List<FeatureModelFeature> elementsToRename = new ArrayList<>();
         element.getContainingFile().accept(new PsiRecursiveElementWalkingVisitor() {
             @Override
             public void visitElement(@NotNull PsiElement e) {
-                if (e instanceof FeatureModelFeature){
-                    if (((FeatureModelFeature) e).getLPQText().contains(element.getFeatureName()) &&
-                        !(lpqToRemove.contains(((FeatureModelFeature) e).getLPQText()))) {
-                        elementsToRename.add((FeatureModelFeature) e);
-                    }
+                if (e instanceof FeatureModelFeature featureModelFeature &&
+                        featureModelFeature.getLPQText().contains(element.getFeatureName()) &&
+                        !(lpqToRemove.contains(featureModelFeature.getLPQText()))) {
+                    elementsToRename.add(featureModelFeature);
                 }
+
                 super.visitElement(e);
             }
         });
@@ -691,7 +680,7 @@ public class FeatureReferenceUtil {
             for (Map.Entry<String, List<PsiReference>> entry : toUpdate.entrySet()) {
                 if (entry.getKey().equals(oldlpqList.get(i))) {
 
-                    for (PsiReference reference:entry.getValue()) {
+                    for (PsiReference reference : entry.getValue()) {
                         reference.handleElementRename(newlpqList.get(i));
                     }
 
@@ -705,7 +694,7 @@ public class FeatureReferenceUtil {
     public static void setElementsToRenameAfterAddingWithChildren(FeatureModelFeature element) {
         Map<String, List<PsiReference>> toUpdate = new HashMap<>();
 
-        List<PsiElement> elementsToUpdate= getElementsToRenameAfterAddingWithChildren(element);
+        List<PsiElement> elementsToUpdate = getElementsToRenameAfterAddingWithChildren(element);
         for (PsiElement e : elementsToUpdate) {
             List<PsiReference> referencedElements = new ArrayList<>();
             for (PsiReference reference : ReferencesSearch.search(e)) {
@@ -717,8 +706,8 @@ public class FeatureReferenceUtil {
     }
 
     private static List<PsiElement> getElementsToRenameAfterAddingWithChildren(FeatureModelFeature element) {
-        List<PsiElement> elementsToUpdate= new ArrayList<>();
-        List<String> lpqToDelete= new ArrayList<>();
+        List<PsiElement> elementsToUpdate = new ArrayList<>();
+        List<String> lpqToDelete = new ArrayList<>();
 
         traverseFeatureWithChildren(element, elementsToUpdate, lpqToDelete);
         return elementsToUpdate;
